Show the replication status after running the replication command
The actual replication work is done asynchronously and it is split
into many subtasks, this patch adds the functionality that traces
every subtask execution, once the task is finished, we can get
replication status of all subtasks. This makes it possible for users
to start the replication ssh command, and wait to see information
about replication progress and errors in their consoles before the
command exits.

For the gerrit triggered replication of receiving a new patchset,
we can also add some features in the future, for example, send
stream events for a finished replication to a node.

Change-Id: I0fc64e93faa9157fc936379de29691282388ba26
(cherry picked from commit 05f71912f4289480533773ffd05cb5cfcbe191bf)
Reviewed-by: Ismo Haataja <ismo.haataja@digia.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>