Support replication of project deletions
The replication plugin can now be configured to listen to project
deletion events and to replicate the project deletions. In the
configuration it can be decided per remote destination whether project
deletions should be replicated to it or not. By default project
deletions are not replicated.

Change-Id: I1207015ed361155dd0353bed86860c794cdb2517
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>
(cherry picked from commit ad923e4e1e4ae3923d15a8b4eee36379423846be)
Reviewed-by: Ismo Haataja <ismo.haataja@digia.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>