Prevent creating repos on extra servers
If using a group to replicate only certain repositories, it is possible
to be in a state where the authGroup is used on some servers but not
others.  If this happened, Gerrit would create the repository on all
servers, even if the authGroup would prevent replicating code to it.
By ensuring the authGroup can see the project first, we don't create
the repository if it's not needed.

Bug: Issue 543
Change-Id: I2f189bdb6007228d0b25a6ce9c0357b3f97e4ced
(cherry picked from commit 7c25cd575572410ef90a53d1af853fde7d011bf8)
Reviewed-by: Ismo Haataja <ismo.haataja@digia.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>